AP_NavEKF3: don't reject setLatLng data timestamped in the future
The check for data more than 5 seconds old used an unsigned
difference, so a timestamp slightly ahead of the EKF's IMU sample
time, which is normal as the clock moves on during a loop, wrapped and
was rejected.  Such timestamps are already clamped to the current time.

+++ libraries/AP_NavEKF3/AP_NavEKF3_PosVelFusion.cpp
@@ -202,8 +202,9 @@ bool NavEKF3_core::setLatLng(const Location &loc, float posAccuracy, uint32_t ti
     setLatLngPosAcc = MAX(posAccuracy, frontend->_gpsHorizPosNoise);
 
     // Correct the position for time delay relative to fusion time horizon assuming a constant velocity
-    // Don't use data more than 5 seconds old
-    if (imuSampleTime_ms - timestamp_ms > 5000) {
+    // Don't use data more than 5 seconds old.  The difference is signed
+    // as a timestamp slightly in the future is normal and is handled below
+    if (int32_t(imuSampleTime_ms - timestamp_ms) > 5000) {
         return false;
     }
